fix(routes): script and googleNLP connected
