+++ app.js
@@ -1,15 +1,18 @@
-// var path = require('path')
-// var bodyParser = require('body-parser')
-var {resolve} = require('path')
-var express = require('express')
-var app = express()
+const bodyParser = require('body-parser')
+const {resolve} = require('path')
+const express = require('express')
+const app = express()
+const router = require('./server/app/routes')
+
 
 
 app.use(require('volleyball'))
 
-// app.use(bodyParser.urlencoded({extended: true}))
-// app.use(bodyParser.json())
+app.use(bodyParser.urlencoded({extended: true}))
+app.use(bodyParser.json())
 
-app.use(express.static(resolve(__dirname, 'phrase-matcher')))
+app.use(express.static(resolve(__dirname, 'browser')))
+
+app.use('/input', router)
 
 app.listen(3000, console.log('Listening at port 3000'))
