+++ browser/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <meta name="viewport" content="width=device-width">
+
+    <title>Phrase matcher</title>
+
+    <link rel="stylesheet" href="style.css">
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+  </head>
+
+  <body>
+    <h1>Phrase matcher</h1>
+    <p>Press the button then say the phrase to test the recognition.</p>
+
+    <button>Start new test</button>
+
+    <div id='main'></div>
+
+    <div>
+        <p class="phrase">Phrase...</p>
+        <p class="output">...diagnostic messages</p>
+    </div>
+
+    <script src="bundle.js"></script>
+  </body>
+</html>
