--- index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
-    <meta name="viewport" content="width=device-width">
-
-    <title>Web Speech API</title>
-
-    <link rel="stylesheet" href="style.css">
-    <!--[if lt IE 9]>
-      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-    <![endif]-->
-  </head>
-
-  <body>
-    <h1>Pick your test</h1>
-    <ul>
-      <li><a href="phrase-matcher/index.html">Phrase matcher</a></li>
-      <li><a href="speak-easy-synthesis/index.html">Speak easy synthesis</a></li>
-      <li><a href="speech-color-changer/index.html">Speech color changer</a></li>
-    </ul>
-
-  </body>
-</html>
