+++ package.json
@@ -5,6 +5,7 @@
   "main": "app.js",
   "dependencies": {
     "@google-cloud/language": "^0.10.3",
+    "axios": "^0.16.1",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
